Using LinkedHashSet to avoid inconsistency
Goose uses a HashSet for iterating topNode candidates
But HashSet doesn't guarantee ordering, so when two candidates have
the same score, the choice is basically random. This is not acceptable.
Now, by using LinkedHashSet we make sure that in case of draw, we choose
the first tag that was found in the DOM tree.